salvando

+++ Algoritmos e Programacao de computadores/ex7/ex7.2.c
@@ -1,10 +1,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
-int areaCirculo(float raio);
+
+double areaCirculo(float raio);
 
 
-int circunferenciaCirculo(float raio);
+double circunferenciaCirculo(float raio);
 
 int main(void){
     float raio;
@@ -12,15 +13,15 @@ int main(void){
     printf("Raio: ");
     scanf("%f", &raio);
 
-    printf("Area = %f\n", areaCirculo(raio));
-    printf("Circunferencia = %f\n", circunferenciaCirculo(raio));
+    printf("Area = %.2f\n", areaCirculo(raio));
+    printf("Circunferencia = %.2f\n", circunferenciaCirculo(raio));
     return 0;
 }
 
-int areaCirculo(float raio){
+double areaCirculo(float raio){
     return M_PI * raio * raio;
 }
 
-int circunferenciaCirculo(float raio){
+double circunferenciaCirculo(float raio){
     return 2 * M_PI * raio;
 }
